Rename to contain sensor context
