+++ src/pages/NewSensor/NewSensorStepThree.js
@@ -25,7 +25,7 @@ import {
 import { goBack } from '../../navigation/actions';
 import { selectNewSensor } from '../../selectors/Entities/sensor';
 import { WizardActions, LocationActions, SensorActions } from '../../actions';
-import { UpdateActions } from '../../actions/Bluetooth/UpdateActions';
+import { SensorUpdateActions } from '../../actions/Bluetooth/SensorUpdateActions';
 import { useLoadingIndicator } from '../../hooks/useLoadingIndicator';
 import { DARKER_GREY, LIGHT_GREY, SUSSOL_ORANGE, WHITE } from '../../globalStyles';
 import { buttonStrings, vaccineStrings } from '../../localization';
@@ -106,7 +106,7 @@ const dispatchToProps = dispatch => {
   const previousTab = () => dispatch(WizardActions.previousTab());
   const exit = () => dispatch(goBack());
   const connectToSensor = sensor => () =>
-    dispatch(UpdateActions.updateSensor(sensor))
+    dispatch(SensorUpdateActions.updateSensor(sensor))
       .then(() => dispatch(SensorActions.createNew()))
       .then(() => {
         ToastAndroid.show(vaccineStrings.sensor_save_success, ToastAndroid.LONG);
