+++ src/pages/VaccinePage.js
@@ -27,7 +27,7 @@ import { buttonStrings } from '../localization';
 import { DARKER_GREY, BLACK, FINALISE_GREEN } from '../globalStyles';
 import { gotoEditSensorPage, gotoFridgeDetailPage, gotoNewSensorPage } from '../navigation/actions';
 import { AfterInteractions } from '../widgets/AfterInteractions';
-import { BlinkActions } from '../actions/Bluetooth/BlinkActions';
+import { SensorBlinkActions } from '../actions/Bluetooth/SensorBlinkActions';
 
 const formatDate = date => moment(date).fromNow();
 const formatTemperature = temperature => `${Math.round(temperature * 10) / 10}°C`;
@@ -187,7 +187,7 @@ const stateToProps = state => {
 };
 
 const dispatchToProps = dispatch => ({
-  blinkSensor: macAddress => dispatch(BlinkActions.startSensorBlink(macAddress)),
+  blinkSensor: macAddress => dispatch(SensorBlinkActions.startSensorBlink(macAddress)),
   toFridgeDetail: fridge => dispatch(gotoFridgeDetailPage(fridge)),
   toEditSensorPage: sensor => dispatch(gotoEditSensorPage(sensor)),
   toNewSensorPage: () => dispatch(gotoNewSensorPage()),
