+++ src/reducers/Bluetooth/index.js
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux';
-import { BlinkReducer } from './BlinkReducer';
-import { ScanReducer } from './ScanReducer';
-import { UpdateReducer } from './UpdateReducer';
+import { SensorBlinkReducer } from './SensorBlinkReducer';
+import { SensorScanReducer } from './SensorScanReducer';
+import { SensorUpdateReducer } from './SensorUpdateReducer';
 
 export const BluetoothReducer = combineReducers({
-  blink: BlinkReducer,
-  scan: ScanReducer,
-  update: UpdateReducer,
+  blink: SensorBlinkReducer,
+  scan: SensorScanReducer,
+  update: SensorUpdateReducer,
 });
